+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/FastLoginBukkit.java
@@ -227,4 +227,8 @@ public class FastLoginBukkit extends JavaPlugin {
         authPlugin = authPluginHook;
         return true;
     }
+
+    public boolean isBungeeCord() {
+        return bungeeCord;
+    }
 }